--- src/resources/ads.json
@@ -1,58 +1,0 @@
-[
-    {
-        "id": 1,
-        "title": "پژو پارس",
-        "price": "25000000",
-        "createdAt": "1402-06-01",
-        "image": "https://media.cgtrader.com/variants/JB8nygcgjtQ2G7WVcRPqGeBV/508c09fe26e63b93baec6d888aa2950650a509b5e04f2f03a4f2af52a04363ea/1.jpg"
-    },
-    {
-        "id": 2,
-        "title": "سمند LX",
-        "price": "29500000",
-        "createdAt": "1402-06-02",
-        "image": "https://preview.free3d.com/img/2018/11/2269257440501433508/ms6k8q3i.jpg"
-    },
-    {
-        "id": 3,
-        "title": "تویوتا کمری",
-        "price": "90000000",
-        "createdAt": "1402-06-03",
-        "image": "https://preview.free3d.com/img/2020/06/2337162504616019433/tyildfpv.jpg"
-    },
-    {
-        "id": 4,
-        "title": "کیا سراتو",
-        "price": "78000000",
-        "createdAt": "1402-06-04",
-        "image": "https://img-new.cgtrader.com/items/2333441/1604e8b81b/large/low-poly-car-3d-model-cartoon-car-3d-model-low-poly-max-obj-3ds-fbx-c4d-dxf.jpg"
-    },
-    {
-        "id": 5,
-        "title": "مزدا 3",
-        "price": "106000000",
-        "createdAt": "1402-06-05",
-        "image": "https://netrinoimages.s3.eu-west-2.amazonaws.com/2019/01/15/575787/225054/cartoon_low_poly_toy_car_3d_model_c4d_max_obj_fbx_ma_lwo_3ds_3dm_stl_2353536.jpg"
-    },
-    {
-        "id": 6,
-        "title": "رنو تندر",
-        "price": "49000000",
-        "createdAt": "1402-06-06",
-        "image": "https://as2.ftcdn.net/v2/jpg/01/83/04/39/1000_F_183043926_r1JkOE0kNtYNm9dOpLpu6wp2v31Qwuwd.jpg"
-    },
-    {
-        "id": 7,
-        "title": "بنز E-Class",
-        "price": "350000000",
-        "createdAt": "1402-06-07",
-        "image": "https://media.sketchfab.com/models/fab332b5caef4811b281eba394b101a9/thumbnails/29317763d1f64064a368d98d656d2629/e43c4a73353f4b5e97d35b1b02cff73b.jpeg"
-    },
-    {
-        "id": 8,
-        "title": "بی‌ام‌و سری 5",
-        "price": "550000000",
-        "createdAt": "1402-06-08",
-        "image": "https://media.sketchfab.com/models/a6e726f298384355bec6ec8b7b720a9e/thumbnails/207f3a3dd5364e5e918b7cb4f37588e9/387a62928c0e4d82a22b390cb0cb3640.jpeg"
-    }
-]
